requirements.txt
absl-py==2.3.0
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.10.0
asttokens==2.2.1
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
bitsandbytes==0.47.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.21.0
comm==0.1.3
confection==0.1.5
contourpy==1.3.2
cut-cross-entropy==25.1.1
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.17.4
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.17.0
einops==0.8.1
en_core_web_sm 
evaluate==0.4.3
exceptiongroup==1.3.0
executing==1.2.0
filelock==3.19.1
fonttools==4.59.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.8
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.23.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.3
jiter==0.10.0
joblib==1.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
Markdown==3.8.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mecab-python3==1.0.10
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
nest-asyncio==1.5.6
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
openai==1.106.1
packaging==25.0
pandas==2.2.3
parso==0.8.3
peft==0.15.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==3.5.1
preshed==3.0.9
prompt-toolkit==3.0.38
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.15.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-crf==0.7.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.0.2
regex==2025.7.34
requests==2.32.5
rich==14.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
sexpdata==1.0.2
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.2
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.8.0+cu126
torchao==0.12.0
torchaudio==2.7.0+cu118
torchvision==0.23.0+cu126
tornado==6.3.1
tqdm==4.67.1
traitlets==5.9.0
transformers==4.30.2
triton==3.4.0
trl==0.21.0
typeguard==4.4.4
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.14.1
tyro==0.9.28
tzdata==2025.2
unsloth 
unsloth_zoo==2025.8.8
uritemplate==4.1.1
urllib3==2.5.0
wandb==0.19.10
wasabi==1.1.3
wcwidth==0.2.6
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.32.post2
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0